requirements.txt
streamlit
python-dotenv
Pillow
langchain-core
langchain-openai
langchain_community
